Flexiple dataset prototype
