+++ fast_llm/data/blended.py
@@ -1,12 +1,12 @@
-import abc
 import logging
 import pathlib
 import time
+import typing
 
 import numpy as np
-import torch.utils.data
 
 from fast_llm.core.distributed import ProcessGroup, safe_barrier
+from fast_llm.data.config import SampledDataset
 from fast_llm.engine.config_utils.run import log_main_rank
 from fast_llm.utils import Assert
 
@@ -18,43 +18,6 @@ except ImportError:
     _extension_available = False
 
 logger = logging.getLogger(__name__)
-
-
-class Dataset(abc.ABC):
-    """
-    A generic dataset class compatible with torch.utils.data.Dataset but with a slightly different signature.
-    """
-
-    @abc.abstractmethod
-    def __getitem__(self, index: int):
-        pass
-
-    @abc.abstractmethod
-    def __len__(self):
-        pass
-
-    @property
-    @abc.abstractmethod
-    def name(self):
-        """
-        A name for the dataset to facilitate identification and debugging.
-        """
-
-
-class RawDataset(Dataset):  # noqa
-    """
-    A raw dataset class containing a list of unsampled, unprocessed samples, i.e., matching what is stored on disk.
-    (Excluding off-line processing prior to training.)
-    Functionally identical to a `Dataset`, but renamed for clarity.
-    """
-
-
-class SampledDataset(Dataset):  # noqa
-    """
-    A sampled dataset class containing a prepared list of samples to be indexed sequentially (as-is) during training.
-    (See the `Sampler` class below.)
-    Functionally identical to a `Dataset`, but renamed for clarity.
-    """
 
 
 class BlendedDataset(SampledDataset):
@@ -73,7 +36,7 @@ class BlendedDataset(SampledDataset):
         name: str = "blended",
         num_samples: int,
         cache_dir: pathlib.Path | None = None,
-        group: ProcessGroup | None = None,
+        group: typing.Optional["ProcessGroup"] = None,
         verbose: bool = True,
         data_sample_warn_time_ms: float = 1000,
     ):
@@ -191,24 +154,3 @@ class BlendedDataset(SampledDataset):
     @property
     def name(self):
         return self._name
-
-
-class Sampler(torch.utils.data.Sampler):
-    """
-    A distributed sampler generating indices for a `SampledDataset` (i.e., the natural numbers).
-    To be used as the `batch_sampler` of a `torch.utils.data.DataLoader`.
-    """
-
-    def __init__(self, total_samples, begin_index, micro_batch_size, data_rank, data_parallel):
-        self._total_samples = total_samples
-        self._begin_index = begin_index
-        self._batch_size = micro_batch_size * data_parallel
-        self._start_idx = data_rank * micro_batch_size
-        self._end_idx = (data_rank + 1) * micro_batch_size
-
-    def __len__(self):
-        return self._total_samples
-
-    def __iter__(self):
-        for idx in range(self._begin_index, self._total_samples - self._batch_size + 1, self._batch_size):
-            yield list(range(idx + self._start_idx, idx + self._end_idx))
