+++ fast_llm/data/config.py
@@ -146,52 +146,38 @@ class AbstractData(abc.ABC):
         pass
 
 
-@config_class()
-class DataConfig(AbstractDataConfig):
+class Dataset(abc.ABC):
     """
-    Configuration for the dataset(s), split and sampling.
-    Currently hard-coded to a GPT dataset.
-    TODO: Extract generalizable content.
+    A generic dataset class compatible with torch.utils.data.Dataset but with a slightly different signature.
     """
 
-    _abstract = False
+    @abc.abstractmethod
+    def __getitem__(self, index: int):
+        pass
 
-    tokenizer: TokenizerConfig = Field(
-        default_factory=TokenizerConfig,
-        desc="Configuration for the tokenizer (for FIM).",
-        hint=FieldHint.feature,
-    )
-    fim: FimConfig = Field(
-        default_factory=FimConfig,
-        desc="Configuration for Fill In the Middle (FIM).",
-        hint=FieldHint.feature,
-    )
-    # TODO: set default to [1,0,0]?
-    split: list[float] = Field(
-        default_factory=lambda: [969, 30, 1],
-        desc="Split ratio for train, valid and test datasets.",
-        hint=FieldHint.core,
-        valid=_validate_split,
-    )
-    format: DatasetSource = Field(
-        default=DatasetSource.list,
-        desc="Format for the dataset definition.",
-        hint=FieldHint.core,
-    )
-    path: list[str] = Field(
-        default_factory=list,
-        desc="Path or list of paths and weights.",
-        hint=FieldHint.core,
-        valid=_validate_path,
-    )
-    data_sample_warn_time_ms: float = Field(
-        default=1000,
-        desc="Warn if a sample takes too long to load.",
-        hint=FieldHint.feature,
-        valid=check_field(Assert.gt, 0),
-    )
-    multiprocessing_context: MultiprocessingContext = Field(
-        default=MultiprocessingContext.spawn,
-        desc="Multiprocessing context. Do not touch.",
-        hint=FieldHint.expert,
-    )
+    @abc.abstractmethod
+    def __len__(self):
+        pass
+
+    @property
+    @abc.abstractmethod
+    def name(self):
+        """
+        A name for the dataset to facilitate identification and debugging.
+        """
+
+
+class RawDataset(Dataset):  # noqa
+    """
+    A raw dataset class containing a list of unsampled, unprocessed samples, i.e., matching what is stored on disk.
+    (Excluding off-line processing prior to training.)
+    Functionally identical to a `Dataset`, but renamed for clarity.
+    """
+
+
+class SampledDataset(Dataset):  # noqa
+    """
+    A sampled dataset class containing a prepared list of samples to be indexed sequentially (as-is) during training.
+    (See the `Sampler` class below.)
+    Functionally identical to a `Dataset`, but renamed for clarity.
+    """
