+++ fast_llm/engine/training/trainer.py
@@ -10,7 +10,7 @@ import torch
 
 from fast_llm.core.distributed import safe_barrier
 from fast_llm.data.config import AbstractData
-from fast_llm.data.data import Data
+from fast_llm.data.gpt.data import GPTData
 from fast_llm.engine.config_utils.run import Run, is_main_rank, log_main_rank, log_pipeline_parallel_main_rank
 from fast_llm.engine.distributed.config import PhaseType
 from fast_llm.engine.distributed.distributed import Distributed
@@ -112,7 +112,7 @@ class Trainer(abc.ABC):
 
     @abc.abstractmethod
     def _get_data(self) -> AbstractData:
-        return Data(
+        return GPTData(
             config=self._config.data,
             distributed_config=self._config.distributed,
             # TODO: `vocab_size` is not generic.
